Trait inexor_rgf_core_plugins::entity_instance_manager::EntityInstanceManager

pub trait EntityInstanceManager: Send + Sync {

    // Required methods
    fn has(&self, id: Uuid) -> bool;
    fn get(&self, id: Uuid) -> Option<Arc<ReactiveEntityInstance>>;
    fn get_by_label(&self, label: &str) -> Option<Arc<ReactiveEntityInstance>>;
    fn get_by_label_with_params(
        &self,
        label: &str
    ) -> Option<(Arc<ReactiveEntityInstance>, HashMap<String, String>)>;
    fn get_all(&self) -> Vec<Arc<ReactiveEntityInstance>>;
    fn get_by_type(&self, ty: &EntityTypeId) -> Vec<Arc<ReactiveEntityInstance>>;
    fn get_ids(&self) -> Vec<Uuid>;
    fn count(&self) -> usize;
    fn count_by_type(&self, ty: &EntityTypeId) -> usize;
    fn count_by_component(&self, component: &ComponentTypeId) -> usize;
    fn count_by_behaviour(&self, behaviour_ty: &BehaviourTypeId) -> usize;
    fn create(
        &self,
        entity_instance: EntityInstance
    ) -> Result<Arc<ReactiveEntityInstance>, EntityInstanceCreationError>;
    fn add_component(
        &self,
        id: Uuid,
        component: &ComponentTypeId
    ) -> Result<(), EntityInstanceComponentAddError>;
    fn remove_component(&self, id: Uuid, component: &ComponentTypeId);
    fn delete(&self, id: Uuid);

}

Required Methods

fn has(&self, id: Uuid) -> bool

Returns true, if an entity instance exists with the given UUID.

fn get(&self, id: Uuid) -> Option<Arc<ReactiveEntityInstance>>

Returns the reactive entity instance with the given UUID or None.

fn get_by_label(&self, label: &str) -> Option<Arc<ReactiveEntityInstance>>

Returns the reactive entity instance with the given label or None.

fn get_by_label_with_params( &self, label: &str ) -> Option<(Arc<ReactiveEntityInstance>, HashMap<String, String>)>

Returns the reactive entity instance and the matched path parameters that matches the given label or None. /org/inexor/local/users/:user_id /org/inexor/local/users/PeterPenacka returns: (instance, {“user_id”: “PeterPenacka”})

fn get_all(&self) -> Vec<Arc<ReactiveEntityInstance>>

Returns all reactive entity instances.

fn get_by_type(&self, ty: &EntityTypeId) -> Vec<Arc<ReactiveEntityInstance>>

Returns all reactive entity instances of the given type.

fn get_ids(&self) -> Vec<Uuid>

Returns all ids.

fn count(&self) -> usize

Returns the count of registered reactive entity instances.

fn count_by_type(&self, ty: &EntityTypeId) -> usize

Returns the count of registered reactive entity instances of the given type.

fn count_by_component(&self, component: &ComponentTypeId) -> usize

Returns the count of registered reactive entity instances which are of the given component.

fn count_by_behaviour(&self, behaviour_ty: &BehaviourTypeId) -> usize

Returns the count of registered reactive entity instances which behaves as the given behaviour.

fn create( &self, entity_instance: EntityInstance ) -> Result<Arc<ReactiveEntityInstance>, EntityInstanceCreationError>

Creates a new reactive entity instance.

fn add_component( &self, id: Uuid, component: &ComponentTypeId ) -> Result<(), EntityInstanceComponentAddError>

Adds the component with the given name to the entity instance with the given id.

fn remove_component(&self, id: Uuid, component: &ComponentTypeId)

Removes the component with the given name from the entity instance with the given id.

fn delete(&self, id: Uuid)

Deletes the reactive entity instance with the given id.