Trait inexor_rgf_core_plugins::entity_type_manager::EntityTypeManager

pub trait EntityTypeManager: Send + Sync {

    // Required methods
    fn get_all(&self) -> Vec<EntityType>;
    fn get_by_namespace(&self, namespace: &str) -> Vec<EntityType>;
    fn has(&self, ty: &EntityTypeId) -> bool;
    fn has_by_type(&self, namespace: &str, name: &str) -> bool;
    fn get(&self, ty: &EntityTypeId) -> Option<EntityType>;
    fn get_by_type(&self, namespace: &str, name: &str) -> Option<EntityType>;
    fn find(&self, search: &str) -> Vec<EntityType>;
    fn count(&self) -> usize;
    fn count_by_namespace(&self, namespace: &str) -> usize;
    fn create(
        &self,
        ty: &EntityTypeId,
        description: &str,
        components: Vec<ComponentTypeId>,
        properties: Vec<PropertyType>,
        extensions: Vec<Extension>
    ) -> Result<EntityType, EntityTypeCreationError>;
    fn add_component(&self, ty: &EntityTypeId, component: &ComponentTypeId);
    fn remove_component(&self, ty: &EntityTypeId, component: &ComponentTypeId);
    fn add_property(&self, ty: &EntityTypeId, property: PropertyType);
    fn remove_property(&self, ty: &EntityTypeId, property_name: &str);
    fn add_extension(&self, ty: &EntityTypeId, extension: Extension);
    fn remove_extension(
        &self,
        entity_ty: &EntityTypeId,
        extension_ty: &ExtensionTypeId
    );
    fn delete(&self, ty: &EntityTypeId);
    fn validate(&self, ty: &EntityTypeId) -> bool;
    fn import(&self, path: &str) -> Result<EntityType, EntityTypeImportError>;
    fn export(&self, ty: &EntityTypeId, path: &str);

}

Required Methods

fn get_all(&self) -> Vec<EntityType>

Returns all entity types.

fn get_by_namespace(&self, namespace: &str) -> Vec<EntityType>

Returns all entity types of the given namespace.

fn has(&self, ty: &EntityTypeId) -> bool

Returns true, if a entity type with the given name exists.

fn has_by_type(&self, namespace: &str, name: &str) -> bool

Returns true, if a entity type with the given fully qualified name exists.

fn get(&self, ty: &EntityTypeId) -> Option<EntityType>

Returns the entity type with the given name or empty.

fn get_by_type(&self, namespace: &str, name: &str) -> Option<EntityType>

Returns the entity type with the given fully qualified name or empty.

fn find(&self, search: &str) -> Vec<EntityType>

Returns all entity types whose names matches the given search string.

fn count(&self) -> usize

Returns the count of entity types.

fn count_by_namespace(&self, namespace: &str) -> usize

Returns the count of entity types of the given namespace.

fn create( &self, ty: &EntityTypeId, description: &str, components: Vec<ComponentTypeId>, properties: Vec<PropertyType>, extensions: Vec<Extension> ) -> Result<EntityType, EntityTypeCreationError>

Creates a new entity type.

fn add_component(&self, ty: &EntityTypeId, component: &ComponentTypeId)

Adds the component with the given component_name to the given entity type.

fn remove_component(&self, ty: &EntityTypeId, component: &ComponentTypeId)

Remove the component with the given component_name from the given entity type.

fn add_property(&self, ty: &EntityTypeId, property: PropertyType)

Adds a property to the given entity type.

fn remove_property(&self, ty: &EntityTypeId, property_name: &str)

Removes the property with the given property_name from the given entity type.

fn add_extension(&self, ty: &EntityTypeId, extension: Extension)

Adds an extension to the given entity type.

fn remove_extension( &self, entity_ty: &EntityTypeId, extension_ty: &ExtensionTypeId )

Removes the extension with the given type from the given entity type.

fn delete(&self, ty: &EntityTypeId)

Deletes the entity type.

fn validate(&self, ty: &EntityTypeId) -> bool

Validates the entity type with the given name. Tests that all components exists.

fn import(&self, path: &str) -> Result<EntityType, EntityTypeImportError>

Imports an entity type from a JSON file file located at the given path.

fn export(&self, ty: &EntityTypeId, path: &str)

Exports the entity type with the given name to a JSON file located at the given path.