Trait inexor_rgf_core_plugins::flow_instance_manager::FlowInstanceManager

pub trait FlowInstanceManager: Send + Sync {
    // Required methods
    fn has(&self, id: Uuid) -> bool;
    fn get(&self, id: Uuid) -> Option<Arc<ReactiveFlowInstance>>;
    fn get_by_label(&self, label: &str) -> Option<Arc<ReactiveFlowInstance>>;
    fn create(
        &self,
        flow_instance: FlowInstance
    ) -> Result<Arc<ReactiveFlowInstance>, FlowInstanceCreationError>;
    fn create_from_type(
        &self,
        ty: &FlowTypeId,
        variables: HashMap<String, Value>,
        properties: HashMap<String, Value>
    ) -> Result<Arc<ReactiveFlowInstance>, FlowInstanceCreationError>;
    fn delete(&self, id: Uuid);
}

Required Methods

fn has(&self, id: Uuid) -> bool

Returns true, if an flow instance exists with the given UUID.

fn get(&self, id: Uuid) -> Option<Arc<ReactiveFlowInstance>>

Returns the flow instance with the given UUID or None.

fn get_by_label(&self, label: &str) -> Option<Arc<ReactiveFlowInstance>>

Returns the flow instance with the given label or None.

fn create( &self, flow_instance: FlowInstance ) -> Result<Arc<ReactiveFlowInstance>, FlowInstanceCreationError>

Creates a new reactive flow instance from the given flow instance descriptor.

The wrapper entity instance will be created as well as entity and relation instances.

All reactive instances will be registered in the ReactiveEntityInstanceManager and the ReactiveRelationInstanceManager.

fn create_from_type( &self, ty: &FlowTypeId, variables: HashMap<String, Value>, properties: HashMap<String, Value> ) -> Result<Arc<ReactiveFlowInstance>, FlowInstanceCreationError>

Create a new reactive flow instance from the flow type by the given name.

It’s possible to individualize the flow instance with templating using the given variables.

The wrapper entity instance will be created as well as entity and relation instances.

All reactive instances will be registered in the ReactiveEntityInstanceManager and the ReactiveRelationInstanceManager.

fn delete(&self, id: Uuid)

Deletes the flow instance with the given id.