Trait inexor_rgf_core_plugins::flow_type_manager::FlowTypeManager

pub trait FlowTypeManager: Send + Sync {

    // Required methods
    fn get_all(&self) -> Vec<FlowType>;
    fn get_by_namespace(&self, namespace: &str) -> Vec<FlowType>;
    fn has(&self, ty: &FlowTypeId) -> bool;
    fn has_by_type(&self, namespace: &str, name: &str) -> bool;
    fn get(&self, ty: &FlowTypeId) -> Option<FlowType>;
    fn get_by_type(&self, namespace: &str, name: &str) -> Option<FlowType>;
    fn find(&self, search: &str) -> Vec<FlowType>;
    fn count(&self) -> usize;
    fn count_by_namespace(&self, namespace: &str) -> usize;
    fn create(
        &self,
        ty: &FlowTypeId,
        description: &str,
        wrapper_entity_instance: EntityInstance,
        entity_instances: Vec<EntityInstance>,
        relation_instances: Vec<RelationInstance>,
        variables: Vec<PropertyType>,
        extensions: Vec<Extension>
    );
    fn add_entity_instance(
        &self,
        ty: &FlowTypeId,
        entity_instance: EntityInstance
    );
    fn update_entity_instance(
        &self,
        ty: &FlowTypeId,
        id: Uuid,
        entity_instance: EntityInstance
    );
    fn remove_entity_instance(&self, ty: &FlowTypeId, id: Uuid);
    fn add_extension(&self, ty: &FlowTypeId, extension: Extension);
    fn update_extension(&self, ty: &FlowTypeId, extension: Extension);
    fn remove_extension(
        &self,
        flow_ty: &FlowTypeId,
        extension_ty: &ExtensionTypeId
    );
    fn add_variable(&self, ty: &FlowTypeId, variable: PropertyType);
    fn update_variable(
        &self,
        ty: &FlowTypeId,
        variable_name: &str,
        variable: PropertyType
    );
    fn remove_variable(&self, ty: &FlowTypeId, variable_name: &str);
    fn delete(&self, ty: &FlowTypeId) -> bool;
    fn validate(&self, ty: &FlowTypeId) -> bool;
    fn import(&self, path: &str);
    fn export(&self, ty: &FlowTypeId, path: &str);

}

Required Methods

fn get_all(&self) -> Vec<FlowType>

Returns all flow types.

fn get_by_namespace(&self, namespace: &str) -> Vec<FlowType>

Returns all flow types.

fn has(&self, ty: &FlowTypeId) -> bool

Returns true, if a flow type with the given name exists.

fn has_by_type(&self, namespace: &str, name: &str) -> bool

Returns true, if a flow type with the given name exists.

fn get(&self, ty: &FlowTypeId) -> Option<FlowType>

Returns the flow type with the given name or empty.

fn get_by_type(&self, namespace: &str, name: &str) -> Option<FlowType>

Returns the flow type with the given name or empty.

fn find(&self, search: &str) -> Vec<FlowType>

Returns all flow types whose names matches the given search string.

fn count(&self) -> usize

Returns the count of flow types.

fn count_by_namespace(&self, namespace: &str) -> usize

Returns the count of flow types of the given namespace.

fn create( &self, ty: &FlowTypeId, description: &str, wrapper_entity_instance: EntityInstance, entity_instances: Vec<EntityInstance>, relation_instances: Vec<RelationInstance>, variables: Vec<PropertyType>, extensions: Vec<Extension> )

Creates a new flow type.

fn add_entity_instance(&self, ty: &FlowTypeId, entity_instance: EntityInstance)

Adds the given entity instance to the flow type with the given name.

fn update_entity_instance( &self, ty: &FlowTypeId, id: Uuid, entity_instance: EntityInstance )

Updates the entity instance with the given id of the flow type with the given name.

fn remove_entity_instance(&self, ty: &FlowTypeId, id: Uuid)

Removes the entity instance with the given id from the flow type with the given name.

fn add_extension(&self, ty: &FlowTypeId, extension: Extension)

Adds the given extension to the given flow type.

fn update_extension(&self, ty: &FlowTypeId, extension: Extension)

Updates the extension with the given type of the given flow type.

fn remove_extension(&self, flow_ty: &FlowTypeId, extension_ty: &ExtensionTypeId)

Removes the extension with the given type from the given flow type.

fn add_variable(&self, ty: &FlowTypeId, variable: PropertyType)

Adds the given variable to the given flow type.

fn update_variable( &self, ty: &FlowTypeId, variable_name: &str, variable: PropertyType )

Updates the variable with the given name of the flow type with the given name.

fn remove_variable(&self, ty: &FlowTypeId, variable_name: &str)

Removes the variable with the given name from the flow type with the given name.

fn delete(&self, ty: &FlowTypeId) -> bool

Deletes the flow type with the given name.

fn validate(&self, ty: &FlowTypeId) -> bool

Validates the flow type with the given name. Tests that all entity types and relation types exists and are valid.

fn import(&self, path: &str)

Imports an flow type from a JSON file file located at the given path.

fn export(&self, ty: &FlowTypeId, path: &str)

Exports the flow type with the given name to a JSON file located at the given path.