Trait inexor_rgf_core_plugins::relation_instance_manager::RelationInstanceManager

pub trait RelationInstanceManager: Send + Sync {

    // Required methods
    fn has(&self, edge_key: &EdgeKey) -> bool;
    fn get(&self, edge_key: &EdgeKey) -> Option<Arc<ReactiveRelationInstance>>;
    fn get_by_outbound_entity(
        &self,
        outbound_entity_id: Uuid
    ) -> Vec<Arc<ReactiveRelationInstance>>;
    fn get_by_inbound_entity(
        &self,
        inbound_entity_id: Uuid
    ) -> Vec<Arc<ReactiveRelationInstance>>;
    fn get_all(&self) -> Vec<Arc<ReactiveRelationInstance>>;
    fn get_by_type(
        &self,
        ty: &RelationTypeId
    ) -> Vec<Arc<ReactiveRelationInstance>>;
    fn get_by_namespace(
        &self,
        namespace: &str
    ) -> Vec<Arc<ReactiveRelationInstance>>;
    fn get_keys(&self) -> Vec<EdgeKey>;
    fn count(&self) -> usize;
    fn count_by_type(&self, ty: &RelationTypeId) -> usize;
    fn count_by_component(&self, component: &ComponentTypeId) -> usize;
    fn count_by_behaviour(&self, behaviour_ty: &BehaviourTypeId) -> usize;
    fn create(
        &self,
        relation_instance: RelationInstance
    ) -> Result<Arc<ReactiveRelationInstance>, RelationInstanceCreationError>;
    fn add_component(
        &self,
        edge_key: &EdgeKey,
        component: &ComponentTypeId
    ) -> Result<(), ReactiveRelationInstanceComponentAddError>;
    fn remove_component(&self, edge_key: &EdgeKey, component: &ComponentTypeId);
    fn delete(&self, edge_key: &EdgeKey) -> bool;

}

Required Methods

fn has(&self, edge_key: &EdgeKey) -> bool

Returns true, if an relation of the given type exists which starts at the given outbound entity and ends at the given inbound entity.

fn get(&self, edge_key: &EdgeKey) -> Option<Arc<ReactiveRelationInstance>>

Returns the ReactiveRelationInstance with the given type_name, which starts at the given outbound entity and ends at the given inbound entity.

fn get_by_outbound_entity( &self, outbound_entity_id: Uuid ) -> Vec<Arc<ReactiveRelationInstance>>

Returns all reactive relation instances of the given outbound entity instance.

fn get_by_inbound_entity( &self, inbound_entity_id: Uuid ) -> Vec<Arc<ReactiveRelationInstance>>

Returns all reactive relation instances of the given inbound entity instance.

fn get_all(&self) -> Vec<Arc<ReactiveRelationInstance>>

Returns all reactive relation instances.

fn get_by_type(&self, ty: &RelationTypeId) -> Vec<Arc<ReactiveRelationInstance>>

Returns all reactive relation instances of the given type.

fn get_by_namespace( &self, namespace: &str ) -> Vec<Arc<ReactiveRelationInstance>>

Returns all reactive relation instances of the given namespace.

fn get_keys(&self) -> Vec<EdgeKey>

Returns all edge keys.

fn count(&self) -> usize

Returns the count of registered reactive relation instances.

fn count_by_type(&self, ty: &RelationTypeId) -> usize

Returns the count of registered reactive relation instances of the given type.

fn count_by_component(&self, component: &ComponentTypeId) -> usize

Returns the count of registered reactive relation instances which are of the given component.

fn count_by_behaviour(&self, behaviour_ty: &BehaviourTypeId) -> usize

Returns the count of registered reactive relation instances which behaves as the given behaviour.

fn create( &self, relation_instance: RelationInstance ) -> Result<Arc<ReactiveRelationInstance>, RelationInstanceCreationError>

Creates a new reactive relation instance.

fn add_component( &self, edge_key: &EdgeKey, component: &ComponentTypeId ) -> Result<(), ReactiveRelationInstanceComponentAddError>

Adds the component with the given name to the relation instance with the given edge key.

fn remove_component(&self, edge_key: &EdgeKey, component: &ComponentTypeId)

Removes the component with the given name from the relation instance with the given edge key.

fn delete(&self, edge_key: &EdgeKey) -> bool

Deletes the reactive relation instance with the given key.